+++ src/blob/mod.rs
@@ -6,7 +6,7 @@ pub(crate) mod index;
 
 pub(crate) use self::config::BlobConfig;
 pub(crate) use self::core::BLOB_INDEX_FILE_EXTENSION;
-pub(crate) use self::core::{Blob, FileName};
+pub(crate) use self::core::{Blob, FileName, DeleteResult};
 pub use self::entry::Entry;
 pub(crate) use self::index::IndexConfig;
 pub(crate) use super::prelude::*;