+++ src/io/unix/iouring.rs
@@ -69,6 +69,10 @@ impl File {
         self.sync.size()
     }
 
+    pub(crate) fn dirty_bytes(&self) -> u64 {
+        self.sync.dirty_bytes()
+    }
+
     pub(crate) async fn write_append_writable_data<R: Send + 'static>(
         &self,
         c: impl WritableDataCreator<R>,
@@ -135,7 +139,10 @@ impl File {
 
     pub(crate) async fn fsyncdata(&self) -> IOResult<()> {
         if let Some(ref rio) = self.rio {
-            rio.fsync(self.sync.std_file_ref()).await
+            let size = self.size();
+            rio.fsync(self.sync.std_file_ref()).await?;
+            self.sync.set_synced_size(size);
+            Ok(())
         } else {
             self.sync.fsyncdata().await
         }
@@ -207,6 +214,9 @@ impl super::super::FileTrait for File {
     }
     fn created_at(&self) -> IOResult<SystemTime> {
         self.created_at()
+    }
+    fn dirty_bytes(&self) -> u64 {
+        self.dirty_bytes()
     }
 
     async fn write_append_writable_data<R: Send + 'static>(&self, c: impl WritableDataCreator<R>) -> IOResult<R> {